Illustrate the sliding operation
In javaslang it is possible to do map operation not only
on a single elements of a stream but also on all the
element of a "window". A window is a gathering of contiguous
elements. Size of the window can be chosen and the sliding
pace too.
Notice how simple operations in javaslang becomes complex
in java vanilla.